refactor(history): split WorkspaceGroupedHistory into sub-modules
Extract the 1308-line monolithic component into grouped-history/:
- types.ts: shared type definitions
- utils/groupingHelpers.ts: timeline grouping and sorting logic
- utils/exportHelpers.ts: file/path/markdown/json export builders
- ConversationRow.tsx: conversation row sub-component
- hooks/useConversations.ts: data fetching and workspace expansion
- hooks/useBatchSelection.ts: batch mode selection state
- hooks/useConversationActions.ts: delete/rename/pin/navigate actions
- hooks/useExport.ts: full export workflow
- index.tsx: main component assembling all hooks

Original file replaced with a single re-export for backward compat.

Also fix conversation modifyTime not updating on user message send
in AcpAgentManager and GeminiAgentManager, causing stale history sort.
